# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Test Laravel REST API",
        "description": "Це тестовий проєкт REST API на Laravel. Документація реалізована за допомогою пакета darkaonline/l5-swagger.",
        "version": "0.2.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Локальний сервер"
        }
    ],
    "paths": {
        "/api/v1/articles": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "Get all articles",
                "operationId": "3fe8bf684063a5479fb27f5432db77e2",
                "responses": {
                    "200": {
                        "description": "Get list of articles",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ArticleResource"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Articles"
                ],
                "summary": "Create new Article",
                "operationId": "7541a17ad3825bab482f62fbb79e6507",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreArticleRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Get one Article by it's ID",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ArticleResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Article not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/articles/{id}": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "Get one Article",
                "operationId": "0e00aae7c370b295281ca65f1d061354",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Article's ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get one Article by it's ID",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ArticleResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Article not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Articles"
                ],
                "summary": "Update existing Article",
                "operationId": "c0ce65fb228f0748dee0abad0396c96d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Article's ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateArticleRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated chosen Article with given data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ArticleResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Article not updated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Articles"
                ],
                "summary": "Delete chosen Article",
                "operationId": "ca3f4f1856466f4b96c1901ff4423c78",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Article's ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Delete Article by it's ID"
                    },
                    "404": {
                        "description": "Article not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/categories": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Get all categories",
                "operationId": "52c3c8d702572c2b4347b0ecf7573e54",
                "responses": {
                    "200": {
                        "description": "Get list of categories",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/CategoryResource"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/categories/{id}": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Get one Category",
                "operationId": "11d2d45398d538612b1a4156d210652e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Category's ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get one Category by it's ID",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CategoryResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Category not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "StoreArticleRequest": {
                "required": [
                    "title",
                    "description",
                    "category_id",
                    "author",
                    "published_at"
                ],
                "properties": {
                    "title": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "Breaking News Title"
                    },
                    "description": {
                        "type": "string",
                        "maxLength": 1024,
                        "example": "This is the content of the news."
                    },
                    "category_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "author": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "John Doe"
                    },
                    "published_at": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-05-19"
                    }
                },
                "type": "object"
            },
            "UpdateArticleRequest": {
                "properties": {
                    "title": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "Breaking News Title"
                    },
                    "description": {
                        "type": "string",
                        "maxLength": 1024,
                        "example": "This is the content of the news."
                    },
                    "category_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "author": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "John Doe"
                    },
                    "published_at": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-05-19"
                    }
                },
                "type": "object"
            },
            "ArticleResource": {
                "title": "Article Resource",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Breaking News Title"
                    },
                    "description": {
                        "type": "string",
                        "example": "This is the content of the news."
                    },
                    "author": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "John Doe"
                    },
                    "category": {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 1
                            },
                            "title": {
                                "type": "string",
                                "example": "Category name"
                            }
                        },
                        "type": "object"
                    },
                    "published_at": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-05-19"
                    }
                },
                "type": "object"
            },
            "CategoryResource": {
                "title": "Category Resource",
                "properties": {
                    "name": {
                        "description": "Tip:\nindex array -> array\nassociative array -> object",
                        "type": "string",
                        "example": "Technology"
                    },
                    "sort_order": {
                        "type": "integer",
                        "example": 1
                    },
                    "show_in_main_menu": {
                        "type": "boolean",
                        "example": true
                    },
                    "parent": {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 1
                            },
                            "name": {
                                "type": "string",
                                "example": "Category name"
                            }
                        },
                        "type": "object"
                    },
                    "children": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "type": "integer",
                                    "example": 1
                                },
                                "name": {
                                    "type": "string",
                                    "example": "Category name"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "articles": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "type": "integer",
                                    "example": 1
                                },
                                "title": {
                                    "type": "string",
                                    "example": "Article title"
                                },
                                "author": {
                                    "type": "string",
                                    "example": "Olaksandr"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "published_at": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-05-19"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Articles",
            "description": "Operations with Articles"
        },
        {
            "name": "Categories",
            "description": "Operations with Categories"
        }
    ]
}